requirements.txt
#django==2.2.6
dj_database_url==0.5.0
whitenoise==3.3.1
gunicorn==19.7.1
#psycopg2==2.7.4

#pillow==6.1.0
django-appconf==1.0.3
#django-compressor==2.2
django-libsass==0.7
libsass==0.19.2
pip
pytz==2019.2
#
setuptools==40.8.0
six==1.12.0
sqlparse==0.3.0
